+++ NEWS.md
@@ -2,6 +2,7 @@
  * new function `launderMunsell()` for fixing encoding of neutral colors
  * new function `validateMunsell()` for checking Munsell color notation (#339)
  * minor bug fix in `col2Munsell()` when passed 0-length character
+ * bug fix in `plotSPC(x, max.depth = 100)` when `x` contains degenerate profiles (no horizons)
  * bug fix in `colorChart()` when using groups and neutral colors
  * bug fix in `soilColorSignature()`, NA incorrectly interpreted when `color` argument is supplied hex encoded sRGB
  * bug fix in `soilColorSignature()`, `cluster::pam()` optimization would very rarely hang